server/requirements.txt
alembic==1.8.0
environs==9.3.1
flasgger==0.9.5
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Limiter==3.3.0
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
flask-openapi3[swagger]==4.0.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.1
gunicorn==20.0.4
jsonschema==3.2.0
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.24.2
numpy==1.26.4
pycryptodome==3.18.0
pydantic==2.9.2
pydantic-extra-types==2.10.1
PyMySQL==1.0.2
pytest==6.2.2
python-dateutil==2.9.0
python-json-logger==2.0.4
requests==2.25.1
ruff==0.6.7
six==1.16.0
SQLAlchemy==1.4.18
sqlalchemy-stubs
uWSGI==2.0.26
boto3==1.35.49 # AWS SDK
boto3-stubs==1.35.49
boto3-stubs[cognito-idp]==1.35.49
botocore==1.35.49
botocore-stubs==1.35.49
phonenumbers==8.13.49    # Python port of Google's `libphonenumber` library
# phonenumberslite # Use this instead of `phonenumbers` if the memory usage is too high.
pyjwt[crypto]==2.9.0
pyhumps==3.8.0
